+++ internal/lsp/workspace.go
@@ -13,7 +13,7 @@ import (
 //
 // https://microsoft.github.io/language-server-protocol/specifications/specification-current/#workspace_workspaceFolders
 func (s *server) workspaceWorkspaceFolders() error {
-	err := s.Send("workspace/workspaceFolders", nil, func(folders *[]protocol.WorkspaceFolder) error {
+	return s.Send("workspace/workspaceFolders", nil, func(folders *[]protocol.WorkspaceFolder) error {
 		s.workspaceMux.Lock()
 		defer s.workspaceMux.Unlock()
 
@@ -29,7 +29,6 @@ func (s *server) workspaceWorkspaceFolders() error {
 		}
 		return nil
 	})
-	return err
 }
 
 // workspace/didChangeWorkspaceFolders